remove __typename as this breaks simple object fields such as Point
if __typename property is present in the mutation we'd get the following error:

```
GraphQL error: Variable "$input" got invalid value { x: 49.405079, y:
9.74352, __typename: "Point" } at "input.patch.coords"; Field
"__typename" is not defined by type PointInput.
```

decided to use this shortcut instead of filtering out all fields not present in the `INPUT_OBJECT` of the
introspection result for every ObjectField.